+++ restserver-example/src/main/java/com/microej/example/HelloEndPoint.java
@@ -20,7 +20,6 @@ public class HelloEndPoint extends RestEndpoint {
 
 	public HelloEndPoint() throws IllegalArgumentException {
 		super(END_POINT);
-
 	}
 
 	@Override